pyproject.toml
[project]
name = "wegpiraten"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "babel>=2.17.0",
    "docx2pdf>=0.1.8",
    "docxtpl>=0.20.1",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "icecream>=2.1.7",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pandas-stubs>=2.3.2.250827",
    "pathlib>=1.0.1",
    "pydantic>=2.11.9",
    "pyinstaller>=6.15.0",
    "pypdf2>=3.0.1",
    "pyyaml>=6.0.2",
    "qrcode[pil]>=8.2",
    "reportlab>=4.4.3",
    "rich>=14.1.0",
    "types-openpyxl>=3.1.5.20250822",
    "types-pyyaml>=6.0.12.20250822",
    "types-qrcode>=8.2.0.20250506",
    "types-reportlab>=4.4.1.20250822",
]
